backend/requirements.txt
# SentinelAI Backend Dependencies
# FastAPI & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Authentication
PyJWT>=2.8.0

# AI/ML - YOLOv8
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python-headless>=4.8.0

# PDF Generation
reportlab>=4.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Environment
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# Utilities
pydantic>=2.0.0
